Refactor Sapling spendAuthSig creation into zcash_primitives::sapling
