+++ zcash_primitives/src/lib.rs
@@ -10,6 +10,7 @@ extern crate sapling_crypto;
 
 use sapling_crypto::jubjub::JubjubBls12;
 
+pub mod sapling;
 mod serialize;
 pub mod transaction;
